Renames and fixes specs in api and specs project to prefix OpenApi module. Gem name to open_api-rswag
