--- rswag-specs/lib/rswag/specs/railtie.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Rswag
-  module Specs
-    class Railtie < ::Rails::Railtie
-      rake_tasks do
-        load File.expand_path('../../tasks/rswag-specs_tasks.rake', __dir__)
-      end
-    end
-  end
-end
