+++ rswag-specs/open_api-rswag-specs.gemspec
@@ -6,9 +6,9 @@ $LOAD_PATH.push File.expand_path('lib', __dir__)
 Gem::Specification.new do |s|
   s.name        = 'rswag-specs'
   s.version     = ENV['TRAVIS_TAG'] || '0.0.0'
-  s.authors     = ['Richie Morris']
+  s.authors     = ['Richie Morris', 'Jay Danielian']
   s.email       = ['domaindrivendev@gmail.com']
-  s.homepage    = 'https://github.com/domaindrivendev/rswag'
+  s.homepage    = 'https://github.com/jdanielian/rswag'
   s.summary     = 'A Swagger-based DSL for rspec-rails & accompanying rake task for generating Swagger files'
   s.description = 'Simplify API integration testing with a succinct rspec DSL and generate Swagger files directly from your rspecs'
   s.license     = 'MIT'
